Multiple External Dataset Addition
Changed the code so it uses Jquery to pull a datafile. The datafile can
contain multiple datasets separated by a double-linebreak. The first
line of each dataset is considered the label. See the sample dataset
for an example.

Because it uses Jquery to fetch a file, it has to be hosted on a web
server to work. It won’t work locally.